[ES|QL] Small refactoring (#213439)
## Summary

Just a small refactoring here to make the code a bit more readable. Also
functions inside functions didnt make a lot of sense there

(cherry picked from commit 5f9f88a80edf8d19aca6713a33ce8ffc70362efa)
