+++ src/platform/packages/shared/kbn-esql-validation-autocomplete/src/definitions/commands.ts
@@ -11,22 +11,12 @@ import { i18n } from '@kbn/i18n';
 import {
   type ESQLColumn,
   type ESQLCommand,
-  type ESQLAstItem,
   type ESQLMessage,
   type ESQLFunction,
   isFunctionExpression,
   isWhereExpression,
-  isFieldExpression,
-  Walker,
 } from '@kbn/esql-ast';
-import {
-  getFunctionDefinition,
-  isAssignment,
-  isColumnItem,
-  isFunctionItem,
-  isFunctionOperatorParam,
-  isLiteralItem,
-} from '../shared/helpers';
+import { isAssignment, isColumnItem, isFunctionItem } from '../shared/helpers';
 import {
   appendSeparatorOption,
   asOption,
@@ -37,7 +27,7 @@ import {
 } from './options';
 import { ENRICH_MODES } from './settings';
 
-import { type CommandDefinition, FunctionDefinitionTypes } from './types';
+import { type CommandDefinition } from './types';
 import { suggest as suggestForSort } from '../autocomplete/commands/sort';
 import { suggest as suggestForKeep } from '../autocomplete/commands/keep';
 import { suggest as suggestForDrop } from '../autocomplete/commands/drop';
@@ -50,6 +40,7 @@ import { suggest as suggestForShow } from '../autocomplete/commands/show';
 import { suggest as suggestForGrok } from '../autocomplete/commands/grok';
 import { suggest as suggestForDissect } from '../autocomplete/commands/dissect';
 import { suggest as suggestForEnrich } from '../autocomplete/commands/enrich';
+import { checkAggExistence, checkFunctionContent } from './commands_helpers';
 
 const statsValidator = (command: ESQLCommand) => {
   const messages: ESQLMessage[] = [];
@@ -82,45 +73,6 @@ const statsValidator = (command: ESQLCommand) => {
     .filter(isFunctionItem);
 
   if (statsArg.length) {
-    function isAggFunction(arg: ESQLAstItem): arg is ESQLFunction {
-      return (
-        isFunctionItem(arg) && getFunctionDefinition(arg.name)?.type === FunctionDefinitionTypes.AGG
-      );
-    }
-    function isOtherFunction(arg: ESQLAstItem): arg is ESQLFunction {
-      return (
-        isFunctionItem(arg) && getFunctionDefinition(arg.name)?.type !== FunctionDefinitionTypes.AGG
-      );
-    }
-
-    function checkAggExistence(arg: ESQLFunction): boolean {
-      if (isWhereExpression(arg)) {
-        return checkAggExistence(arg.args[0] as ESQLFunction);
-      }
-
-      if (isFieldExpression(arg)) {
-        const agg = arg.args[1];
-        const firstFunction = Walker.match(agg, { type: 'function' });
-
-        if (!firstFunction) {
-          return false;
-        }
-
-        return checkAggExistence(firstFunction as ESQLFunction);
-      }
-
-      // TODO the grouping function check may not
-      // hold true for all future cases
-      if (isAggFunction(arg) || isFunctionOperatorParam(arg)) {
-        return true;
-      }
-
-      if (isOtherFunction(arg)) {
-        return (arg as ESQLFunction).args.filter(isFunctionItem).some(checkAggExistence);
-      }
-
-      return false;
-    }
     // first check: is there an agg function somewhere?
     const noAggsExpressions = statsArg.filter((arg) => !checkAggExistence(arg));
 
@@ -144,26 +96,6 @@ const statsValidator = (command: ESQLCommand) => {
         }))
       );
     } else {
-      function isConstantOrAggFn(arg: ESQLAstItem): boolean {
-        return isLiteralItem(arg) || isAggFunction(arg);
-      }
-      // now check that:
-      // * the agg function is at root level
-      // * or if it's a operators function, then all operands are agg functions or literals
-      // * or if it's a eval function then all arguments are agg functions or literals
-      // * or if a named param is used
-      function checkFunctionContent(arg: ESQLFunction) {
-        // TODO the grouping function check may not
-        // hold true for all future cases
-        if (isAggFunction(arg) || isFunctionOperatorParam(arg)) {
-          return true;
-        }
-        return (arg as ESQLFunction).args.every(
-          (subArg): boolean =>
-            isConstantOrAggFn(subArg) ||
-            (isOtherFunction(subArg) ? checkFunctionContent(subArg) : false)
-        );
-      }
       // @TODO: improve here the check to get the last instance of the invalidExpression
       // to provide a better location for the error message
       // i.e. STATS round(round(round( a + sum(b) )))
